Fix logic ChargingSession

+++ ev-charging-backend/src/main/java/com/ev/evchargingsystem/model/response/CarResponse.java
@@ -7,6 +7,6 @@ public class CarResponse {
     private Integer id;
     private String brand;
     private String color;
-    private Integer initBattery;
+    private double initBattery;
     private String licensePlate;
 }
